bitcode & llvm arm build fixes
introduce a mkllvmcmds variable that can be on or off even if llvm is
used while crossbuilding.

+++ minix/servers/is/dmp_rs.c
@@ -45,7 +45,8 @@ void rproc_dmp()
 	printf("%13s %9d %5d %6s %4d %4ld %8lu %5dx %s",
   		rpub->label, rpub->endpoint, rp->r_pid,
 		s_flags_str(rp->r_flags, rpub->sys_flags), rpub->dev_nr,
-		rp->r_period, rp->r_alive_tm, rp->r_restarts,
+		(unsigned long) rp->r_period,
+		(unsigned int) rp->r_alive_tm, rp->r_restarts,
 		rp->r_args
   	);
 	printf("\n");
